src/lambda_funcs/4_viz_gen/lawn_viz_gen/requirements.txt
pandas==1.1.2
boto3
sqlalchemy
psycopg2-binary==2.8.6
plotly==4.10.0